Data Analyst Scope or Work (SOW )
In this activity, I got practical experience developing a Scope of Work document with the help of a handy template. I then completed (SOW) an example SOW for an imaginary project of my choosing and I learned how analysts outline the work they are going to perform. By the time I completed this project, I was familiar with an essential, industry-standard tool, and gained comfort asking the right questions to develop an SOW.